修复 C 语言实现问题

+++ base/random_number/c/linear_congruential_generator.c
@@ -61,12 +61,12 @@ Xₙ₊₁ = (A × Xₙ + B) Mod M
 void main(){
 	long A=25214903917;
 	int B=11;
-	long M=1<<48;
-	int X=2;
+	long M=1L<<48;
+	long X=2;
 	// X是随机数种子
 	int i;
 	for(i=0;i<11;i++){
 		X = (A * X + B) % M;
-		printf("%d\n",X);
+		printf("%ld\n",X);
 	}
 }
